+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Begin to categorize the datasets in the package with `@concept` tags
 * Begin to flesh out more general conception of CDA
+* `datasets()`: added `ndim` arg; package="vcdExtra" default
 
 ## Version 0.8.7 (2025-12-10)